move timestamps to common package
