+++ apps/web/package.json
@@ -31,6 +31,7 @@
 		"test:unit": "vitest run"
 	},
 	"dependencies": {
+		"@ficus/common": "workspace:*",
 		"@iconify/svelte": "catalog:",
 		"@libsql/client": "catalog:",
 		"@nick/clsx": "catalog:",
