+++ packages/common/tsconfig.json
@@ -6,5 +6,6 @@
 		"sourceMap": true,
 		"strict": true
 		// "moduleResolution": "bundler",
-	}
+	},
+	"include": ["src"]
 }
